app components dropdown-menu: add dropdown menu component
Wrap some of the DropdownMenu components with our styles.
